# api/openapi.yaml
info:
  title: PaySuper payment solution service.
  version: "1.0.0"
  description: Swagger Specification for PaySuper Management API.
servers:
  - url: "https://api.pay.super.com"
    description: Production API
components:
  schemas:
...
    CountriesList:
      type: object
      properties:
        countries:
          type: array
          items:
            properties:
              iso_code_a2:
                description: "Two-letter country code in ISO 3166, in uppercase (for instance US)."
                type: string
              region:
                description: "A PaySuper system's relative geographical area for this country."
                type: string
              currency:
                description: "The currency of the country. Three-letter Currency Code ISO 4217, in uppercase."
                type: string
              payments_allowed:
                description: "Has a true value if a payment is allowed from this country."
                type: boolean
              change_allowed:
                description: "Has a true value if the customer can select another country in case of the payments are disallowed for this country."
                type: boolean
              vat_enabled:
                description: "Has a true value if VAT is calculated and paid for this country."
                type: boolean
              vat_currency:
                description: "VAT currency. It can differ from the country currency. Three-letter Currency Code ISO 4217, in uppercase."
                type: string
              price_group_id:
                description: "The unique identifier for the price group this country belongs to."
                type: string
              vat_threshold:
                type: object
                properties:
                  year:
                    description: "The maximum value of the PaySuper's revenue (since the beginning of the year) from this country which subjects to VAT."
                    type: number
                  world:
                    description: "The maximum value of the PaySuper's worldwide revenue (since the beginning of the year) which subjects to VAT."
                    type: number
              vat_period_month:
                description: "The length of the taxation period in months."
                type: integer
              vat_deadline_days:
                description: "The number of days after the taxation period end to pay VAT."
                type: integer
              vat_store_years:
                description: "The number of years to store the VAT payment history for this country."
                type: integer
              vat_currency_rates_policy:
                description: "The VAT calculation policy with a currency that differs from this country's VAT currency. Available values: on-day (the conversion rate on the day of payment), last-day (the conversion rate on the last day of a tax period), mid-month (the system reserved option of a conversion rate calculation. This option is disabled.)."
                type: string
              vat_currency_rates_source:
                description: "The source of currency rates to calculate the tax amount."
                type: string
              payer_tariff_region:
                description: "A PaySuper system's relative geographical area for this country. Available values: asia, europe, latin_america, russia_and_cis, worldwide."
                type: string
              high_risk_payments_allowed:
                description: "Has a true value if payments from this country are allowed. Has a false value for a high-risk merchant."
                type: boolean
              high_risk_change_allowed:
                description: "Has a true value if the customer can select another country while the payments are disallowed for a country determined by their IP address."
                type: boolean
...
    Country:
      type: object
      properties:
        iso_code_a2:
          description: "Two-letter country code in ISO 3166, in uppercase (for instance US)."
          type: string
        region:
          description: "A PaySuper system's relative geographical area for this country."
          type: string
        currency:
          description: "The currency of the country. Three-letter Currency Code ISO 4217, in uppercase."
          type: string
        payments_allowed:
          description: "Has a true value if a payment is allowed from this country."
          type: boolean
        change_allowed:
          description: "Has a true value if the customer can select another country in case of the payments are disallowed for this country."
          type: boolean
        vat_enabled:
          description: "Has a true value if VAT is calculated and paid for this country."
          type: boolean
        vat_currency:
          description: "VAT currency. It can differ from the country currency. Three-letter Currency Code ISO 4217, in uppercase."
          type: string
        price_group_id:
          description: "The unique identifier for the price group this country belongs to."
          type: string
        vat_threshold:
          type: object
          properties:
            year:
              description: "The maximum value of the PaySuper's revenue (since the beginning of the year) from this country which subjects to VAT."
              type: number
            world:
              description: "The maximum value of the PaySuper's worldwide revenue (since the beginning of the year) which subjects to VAT."
              type: number
        vat_period_month:
          description: "The length of the taxation period in months."
          type: integer
        vat_deadline_days:
          description: "The number of days after the taxation period end to pay VAT."
          type: integer
        vat_store_years:
          description: "The number of years to store the VAT payment history for this country."
          type: integer
        vat_currency_rates_policy:
          description: "The VAT calculation policy with a currency that differs from this country's VAT currency. Available values: on-day (the conversion rate on the day of payment), last-day (the conversion rate on the last day of a tax period), mid-month (the system reserved option of a conversion rate calculation. This option is disabled.)."
          type: string
        vat_currency_rates_source:
          description: "The source of currency rates to calculate the tax amount."
          type: string
        payer_tariff_region:
          description: "A PaySuper system's relative geographical area for this country. Available values: asia, europe, latin_america, russia_and_cis, worldwide."
          type: string
        high_risk_payments_allowed:
          description: "Has a true value if payments from this country are allowed. Has a false value for a high-risk merchant."
          type: boolean
        high_risk_change_allowed:
          description: "Has a true value if the customer can select another country while the payments are disallowed for a country determined by their IP address."
          type: boolean
...
    ResponseErrorMessage:
      type: object
      properties:
        code:
          description: "The error code."
          type: string
        message:
          description: "The error short description."
          type: string
        details:
          description: "The error details."
          type: string
    Key:
      type: object
      properties:
        id:
          description: "@inject_tag: json\" i\" bson\"_i\""
          type: string
        code:
          description: "@inject_tag: json\" cod\" validate\"required,max=5\""
          type: string
        key_product_id:
          description: "@inject_tag: validate\" required,hexadecimal,len=2\""
          type: string
        platform_id:
          description: "@inject_tag: validate\" required,max=25\""
          type: string
        order_id:
          description: "@inject_tag: validate\" omitempty,hexadecimal,len=2\""
          type: string
        created_at:
          type: object
          properties:
            seconds:
              description: "9999-12-31T23:59:59Z inclusive."
              type: integer
            nanos:
              description: "inclusive."
              type: integer
        reserved_to:
          type: object
          properties:
            seconds:
              description: "9999-12-31T23:59:59Z inclusive."
              type: integer
            nanos:
              description: "inclusive."
              type: integer
        redeemed_at:
          type: object
          properties:
            seconds:
              description: "9999-12-31T23:59:59Z inclusive."
              type: integer
            nanos:
              description: "inclusive."
              type: integer
...
paths:
  /api/v1/country:
    get:
      summary: Get the list of currencies
      operationId: countryGet
      description: "Get the full list of currencies using the country's name"
      tags:
      - Country
      parameters:
      - name: name
        description: "The country's name or two-letter country code by ISO 3166-1."
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Returns the list of currencies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountriesList"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
  /api/v1/country/{code}:
    get:
      summary: Get the country data
      operationId: countryGetById
      description: "Get the country data by two-letter country code in ISO 3166-1"
      tags:
      - Country
      parameters:
      - name: code
        description: "Two-letter country code in ISO 3166-1"
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "Returns the country data (taxes, currency, region and others)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Country"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
  /admin/api/v1/keys/{key_id}:
    get:
      summary: "Get the key's data"
      operationId: keysIdPathGetKeyInfo
      description: "Get the key's data"
      tags:
      - Key
      - Onboarding
      parameters:
      - name: key_id
        description: The unique identifier for the key.
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "Returns the key's data"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Key"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
...
openapi: "3.0.2"